Move `ExposureType` to dbt/artifacts
