+++ core/dbt/artifacts/resources/v1/exposure.py
@@ -1,0 +1,9 @@
+from dbt_common.dataclass_schema import StrEnum
+
+
+class ExposureType(StrEnum):
+    Dashboard = "dashboard"
+    Notebook = "notebook"
+    Analysis = "analysis"
+    ML = "ml"
+    Application = "application"
