+++ tests/unit/test_contracts_graph_unparsed.py
@@ -20,12 +20,12 @@ from dbt.contracts.graph.unparsed import (
     UnparsedExposure,
     MaturityType,
     Owner,
-    ExposureType,
     UnparsedMetric,
     UnparsedMetricTypeParams,
     UnparsedMetricInputMeasure,
     UnparsedVersion,
 )
+from dbt.artifacts.resources import ExposureType
 from dbt.artifacts.schemas.results import FreshnessStatus
 from dbt.node_types import NodeType
 from .utils import ContractTestCase
